% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qctimeseries-package.R
\docType{package}
\name{qctimeseries}
\alias{qctimeseries-package}
\alias{qctimeseries}
\title{qctimeseries: Window-by-window QC for large time series}
\description{
Tools to add/remove QC flag columns, review data in an interactive Shiny/plotly app,
summarize progress, and apply flags to produce cleaned series.
}
\section{Vignette}{

A \emph{vignette} is a long-form tutorial with runnable examples.
\strong{This is the user manual for qctimeseries.}

\preformatted{
  vignette("overall_workflow", package = "qctimeseries")
}

If you installed without vignettes, the vignette can be (temporarily) found at:

\preformatted{
  https://crustnet.org/qctimeseries/
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/anthonydn/qctimeseries}
  \item Report bugs at \url{https://github.com/anthonydn/qctimeseries/issues}
}

}
\author{
\strong{Maintainer}: Anthony Darrouzet-Nardi \email{adarrouzetnardi@gmail.com}

}
